Crate alloy_eip2124

EIP-2124 implementation based on https://eips.ethereum.org/EIPS/eip-2124.

Previously version of Apache licenced ethereum-forkid.

Structs

EnrForkIdEntry

Represents a forward-compatible ENR entry for including the forkid in a node record via EIP-868. Forward compatibility is achieved via EIP-8.

ForkFilter

Filter that describes the state of blockchain and can be used to check incoming ForkIds for compatibility.

ForkHash

CRC32 hash of all previous forks starting from genesis block.

ForkId

A fork identifier as defined by EIP-2124. Serves as the chain compatibility identifier.

ForkTransition

Represents a transition from one fork to another

Head

Describes the current head block.

Enums

ForkFilterKey

How to filter forks.

ValidationError

Reason for rejecting provided ForkId.